introduce internal `useRender` hook
We already use the `render` function helper in all of our components,
but we also need _some_ way of storing persistent functions (such as the
`mergeRefsFn`).

With this hook, we can embed it in the hook and we can ensure it's
always stable. We also won't run into an issue where we accidentally
forget something because it will be shipped by default.