fix crash with TanStack Virtualizer
We already verify that the `activeOptionIndex` is inside the `options`
before we call the `virtualizer.scrollToIndex` function.

However, if you do this when the virtualizer isn't ready internally yet,
then you get an error like `unexpected undefined`.

Found an open issue on their GitHub: https://github.com/TanStack/virtual/issues/879

Don't know if there is a better fix we can apply here, but delaying the
`scrollToIndex` does seem to fix the problem.

Another thing I tried is waiting for the DOM to be ready, but we already
only run this code _if_ the DOM `el` is available.