`opening` and `closing` should keep the hierarchy into account
The `opening` and `closing` state was based on the transition data of
the *current* element. This means that if the parent stops transitioning
before children that the `Opening` or `Closing` state is also gone.

But should should be based on the full hierarchy such that the whole
`Dialog` is still closing if the `Dialog` or any of its children is
transitioning out.