reduce confusion
Inlines the single `isActive` call, and renames the `isActiveOption` to
`isActive` such that the API is similar to other components.

Unfortunately, due to the virtualization layer, not all `ComboboxOption`
IDs are known/available, so instead we also use the `value` itself to
lookup the active index by comparing the `value` values.