patch `@tanstack/virtual-core`
There is a timing issue if the Virtualizer is enabled and disabled in
rapid succession. This happens in our combobox when you see a list of
items and quickly search for something which in turn filters the list
and the options will become empty.

The `scrollToIndex` checks if an element's DOM node is available (via
the measurementsCache) but it also has an internal `setTimeout` where
the same code is used, but it's not re-checking the measurementsCache
cache.

The measurementsCache is also cleared the moment the `enabled` prop
becomes false. This means that there will be a split second where the
`measurementsCache[x]` will be `undefined` even though it was not
undefined moments ago.

In this scenario, it's safe to just not scroll (yet), but instead the
`@tanstack/virtual-core` is throwing an error via the `notUndefined`
function.

We unfortunately can't catch the error because it happens in a
`setTimeout`...

Instead of throwing, this patch ignores the result if it becomes
`undefined`. This also means that we don't scroll to the item, but there
is no item to scroll to in the first place.