Fix incorrect double invocation of menu items (#3766)
This PR fixes an issue where menu items would be invoked twice when you
click on an item.

This is due to a bug in the `useQuickRelease` hook where the timestamp
was not reset correctly, which means that the actual click caused a
quick release trigger. This in turn selects / invokes the menu item even
though it was already triggered by the click of the user.

This PR fixes that by resetting the timestamp as soon as possible, which
also cleans up the code a bit because there is only 1 branch where we
need to reset instead of in every branch before returning.

Fixes: #3763
Fixes: #3764
Fixes: #3765