Ensure `buttonRef.current.click()` works (#3768)
This PR fixes an issue where adding a ref to the `<MenuButton
ref={btnRef}>` and later calling `btnRef.current.click()` would not open
the `Menu`.

This is happening because recently we started using `pointerdown`
instead of `click` to open the `Menu`. So the only way to open the
`Menu` programmatically is to call `btnRef.current.dispatchEvent(new
PointerEvent('pointerdown', { bubbles: true }))` which is a bit of a
mouthful.

We also recently fixed an issue where the `Listbox` would immediately
close after opening the Listbox on touch devices. That solution required
us to not only handle the `pointerdown` event but also the `click`
event.

So if anything, this PR makes the code more consistent between the
`Menu` and `Listbox` component behavior and in turn solves this
`ref.current.click()` issue.

This PR also does some internal refactoring to make the code a bit
cleaner.

Fixes: #3749